+++ changelog.md
@@ -1,6 +1,7 @@
 # PHPMailer Change Log
 
 ## WIP
+* Add partial support for [RFC6530 SMTPUTF8](https://www.rfc-editor.org/rfc/rfc6530), permitting Unicode characters in local parts of addresses, thanks to @arnt and ICANN
 * Use rfc-editor.org for all RFC docs links
 
 ## Version 6.9.2 (October 9th, 2024)